wdio-cli: add test for missingConfigurationPrompt
